refactor(task-manager): retire PrefectTask and finalize package layout
Remove the PrefectTask god class, split the flat package into flow_run and event sub-packages, and rewire the CLI and event resolver onto the new modules.
