+++ backend/infrahub/cli/tasks.py
@@ -7,7 +7,7 @@ from prefect.client.schemas.objects import StateType
 
 from infrahub import config
 from infrahub.services.adapters.workflow.worker import WorkflowWorkerExecution
-from infrahub.task_manager.task import PrefectTask
+from infrahub.task_manager.flow_run.retention import FlowRunRetention, PrefectFlowRunRetentionClient
 from infrahub.tasks.dummy import DUMMY_FLOW, DummyInput
 from infrahub.workers.dependencies import build_tls_registry
 from infrahub.workflows.initialization import setup_task_manager
@@ -74,10 +74,11 @@ async def flow_runs(
 
     config.load_and_exit(config_file_name=config_file)
 
-    await PrefectTask.delete_flow_runs(
-        days_to_keep=days_to_keep,
-        batch_size=batch_size,
-    )
+    async with get_client(sync_client=False) as client:
+        await FlowRunRetention(client=PrefectFlowRunRetentionClient(client)).purge(
+            days_to_keep=days_to_keep,
+            batch_size=batch_size,
+        )
 
 
 @flush_app.command()
@@ -94,6 +95,7 @@ async def stale_runs(
 
     config.load_and_exit(config_file_name=config_file)
 
-    await PrefectTask.delete_flow_runs(
-        states=[StateType.RUNNING], delete=False, days_to_keep=days_to_keep, batch_size=batch_size
-    )
+    async with get_client(sync_client=False) as client:
+        await FlowRunRetention(client=PrefectFlowRunRetentionClient(client)).purge(
+            states=[StateType.RUNNING], delete=False, days_to_keep=days_to_keep, batch_size=batch_size
+        )
